Don't use pop-to-buffer-same-window. Fixes #14.
I hadn't noticed this problem because my .emacs sizes the frame just
wide enough for ~80 char columns side-by-side, and does a `C-x 3` split
window.

As for the `pop-to-buffer-same-window`? That was a remnant of old code
that tried to keep the REPL in the same bottom-right window in a 3
column layout. Although I'd abandoned that (and anyway it wouldn't be
appropriate in a `racket-mode` intended for use by other people), I
didn't abandon it completely: This was still left over.