Implement racket-help using raco doc.
Immediate motivation is that I don't have the ,doc metacommand working
for all langs. For example it doesn't work for Typed Racket.

Admittedly it would be good for that metacommand to work for all langs,
so you could type `,doc foo` in the REPL. I do want to figure that
out. XREPL can do it. I'm frustrated I can't (yet).

However in the meantime, the Emacs command per se may as well simply
shell out to raco doc. It always works.