Graceful error on attempt to require racket/gui/base.
Although this is not yet a resolution for issue #26, it at least makes
it clear that racket-mode doesn't yet support GUI programs. It avoids
people getting into the state where a GUI app is "frozen", and/or where
running again gives the "cannot instantiate `racket/gui/base' a second
time in the same process" error or even segfaults.