Don't use racket/sandbox. Do support GUI programs.
In this commit, we always require racket/gui/base, even for non-GUI user
programs. Doing so is annoying on a GUI OS, and AFAIK a fail on headless
Unix like a Linux VPS. Therefore loading r/g/b on-demand remains to be
done in a future commit.

Meanwhile, this is the first commit where running GUI programs works
properly. The key is to run read-eval-print-loop from the thread
associated with the eventspace. That way,
current-get-interaction-input-port works as-designed, providing an input
port that will continue to pump GUI events while waiting to read.