Interim commit with conditional text/gui.
The mechanics to work whether racket/gui/base is required or not.

But not yet the mechanism to load r/g/b just once on-demand in the main
custodian, and reuse it in additional runs (to avoid the "can't
instantiate r/g/b twice in the same process" error message and possible
segfault).