Checkpoint.
Use `raise` to escape REPL thread. In the handler (still in the REPL
thread), do `channel-put` to main thread then exit REPL thread.

Change our module-name-resolver to use this when it catches a require
r/g/b at the REPL prompt. Instead of an error, now it can do the
require! However it needs to restart the REPL for that module.

Hopefully the next commit will be using that same mechanism generally --
to catch require r/g/b when loading a module, requiring r/g/b in the
main thread/custodian/namespace, then restarting the module load -- as a
replacement for the imports-gui? approach.