Fix Racket indentation.
In commit acef151 I misunderstood the byte compiler warning and resolved
it the wrong way. We do need to use the calculate-lisp-indent-last-sexp
variable from lisp-mode -- because we're using its calculate-lisp-indent
function via using its lisp-indent-line function.

So the answer wasn't to create our own variable. Instead we simply need
to (defvar calculate-lisp-indent-last-sexp) -- thereby telling the byte
compiler we're aware it's a global variable and we're using it
intentionally.