Add find-signature and use it in ,type
find-signature is like find-definition, but instead of returning the
location of the definition, it returns a string version of the
s-expression for the function's `(name args ...)`.

Now, when ,type cannot find a TR type, nor a contract, it can try to at
least return a signature.

For the future: It would be neat to have a `find-bluebox`, which returns
something like the classic "blue boxes" in Racket documentation. That
is, first the signature, then a list of the argument names again with
the contract (or type) of each, and finally the contract (or type) of
the return value.