Display Racket `#<pict>` values as images in the REPL.
As acknowledged in THANKS.md, this draws very heavily on the pull
request @gcr made to Geiser:

  https://github.com/jaor/geiser/pull/1

One visible difference is that I'm using plain print for the non-image
case, not pretty-print, because the latter seems too big a change from
the status quo. Maybe in some future commit that could be an option.