Make defn.rkt tests work on Racket 6.0 and earlier.
Apparently `raco test` in Racket > 6.0 does the equivalent of
`(current-namespace (module->namespace <module-being-tested>))`.

Our `find-definition` relies on the ability to convert a string or
symbol to fully-enriched identifier syntax using
`namespace-symbol->identifier`. (Remember that the usage of this is an
Emacs mode where we are getting text, not Racket syntax.)