Don't indent whole-line single semicolon comments specially.
A lot of Racket code -- especially that written in DrRacket -- uses
single semicolons (instead of double semicolons) for whole-line
comments.

By contrast the Emacs tradition uses single semicolons exclusively for
end-of-line comments; `lisp-indent-line' insists on always indenting
them as end-of-line, even when they're the only thing on a line. Let's
stop doing that.

We do continue to support M-; inserting a single semicolon end-of-line
comment, when in a non-blank line.