C-a should do comint-bol. Fixes #87.
`comint-mode' has two ways of defining the prompt. Apparently when
`comint-use-prompt-regexp' is nil, a field is automatically inserted
that makes the usual `beginning-of-line' do the right thing. However
racket-repl-mode uses non-nil `comint-use-prompt-regexp'. In that case
it seems the simplest thing to do is bind C-a to the provided
`comint-bol', which behaves as desired.