Help should always try to search.
Commit 680b1b6 dropped a test for "Sending to web browser" up in Elisp,
else retry using search. Restore that down in cmds.rkt.

Need to cover the case where `find-help` does not raise an exception but
instead prints "No current binding, but provided by: ...". In that case
we want to retry using `search-for`, just like in the case where
`find-help` raises an exception because the term isn't provided by
anything.