Use stricter comint-prompt-regexp. Fixes #114.
                  Output vs. prompt. Fight!

                      Sentinels Suck

Yes, it's another exciting chapter in the saga of comint-mode and
prompts.

Although commit 680b1b6 means that we're no longer using
`comint-redirect-send-command-to-process' to capture command output, and
therefore comint-prompt-regexp is no longer quite as critical -- it can
still cause bugs.

In this case it was a hilarious combo of a weak regexp and a comint-mode
feature that helpfully deletes redundant prompts. Which it believed many
of the "> " strings output by racket/trace to be. So it deleted them.

This time around? Try using \uFEFF "ZERO WIDTH NON-BREAKING SPACE".

And, do use beginning of line in the regexp. (Don't bother trying to use
an unusual char to prefix the prompt. You might think that would help
with the case where e.g. a `display` instead of `displayln` means that
the program output runs into the next prompt, and hey we could handle
that. But it doesn't work, because comint-mode has assumptions about
beginning of line, regardless of the regexp.)

Will this be the last battle in the war? I hope so. I doubt it.