Use exn:break-continuation in uncaught-exception-handler.
`with-handlers` installs a continuation barrier that prevents
exn:break-continuation from being usuable in an
uncaught-exception-handler.

Taking that into account, exn:break handling now works.

However, on the next racket-run, we hang. Based on printfs, the hang
seems to be in channel.rkt, in put/stop, on the channel-put. I can't
figure out why.

I'm committing this as a check-point.