Simpler company-mode interop and faster capf
Use the approach where company-mode will call our standard
completion-at-point-function and look for some additional properties to
be returned.

This is much simpler and should be more reliable.

Also, use `completion-table-dynamic' as suggested by @dgutov.
(Don't bother to check for and maybe use `completion-table-with-cache',
new in 24.4. It helps by wrapping/caching a slow operation like talking
to a subprocess. But we don't need it because we're already caching to
minimize that: See `racket--get-namespace-symbols' and
`racket--invalidate-completion-cache'.)