Support font-lock-maximum-decoration with 4 levels
Closes #139.

Generate lists for Racket keywords, builtins, and types.

The question of what is a "keyword" and a "builtin" is not so simple
in Racket:

1. The distinction between the two is squishy, and from one point of
   view Racket has 1400+ "primitives" (!).

2. As for "builtins", there are many, many "batteries included"
   libraries in the main distribution. Where to draw the line?

3. More fundamentally, Racket is a language for making languages.
   Ultimately the only way to be 100% correct is to do something "live"
   with namespace-mapped-symbols. But I don't see that as performant for
   Emacs font-lock.

Here I'm saying that:

(a) "keywords" are syntax (only) from racket/base

(b) "builtins" are everything else provided by #lang racket and #lang
    typed/racket (except the capitalized Types from typed/racket go into
    their own list). Plus for modern macros, racket/syntax and a few
    items from syntax/parse (but not all the syntax classes, because
    things like `id` and `str` are too "generic" and too likely to be
    user program identifiers).

Is that somewhat arbitrary? Hell yes. It's my least-worst, practical
idea for now. Also, IMHO it's an improvement over getting pull
requests to add people's favorites, a few at a time. At least this way
is consistent, and can be regenerated programatically as Racket
evolves.