sysms: Sort namespace symbols
Now that commit c737789 exposes the symbols directly to the user via
`ido-completing-read', sort it, for cosmetic reasons.

(Potentially sorting could be for performance reasons, too. Particularly
if we printed an Elisp array instead of a list, and tweaked completion
to do a binary search? But performance is already fine, as far as I can
tell. The motivation now is only cosmetic.)