Set fill-column back to default 70.
Although I like it at 78, it's weird to edit files where the comments
were formatted at 70. Doing so to existing comments introduces "noise"
to diffs.

Instead, it would be better to do a single "code style" commit: One that
changes fill-column to 78, and also reformats all the existing comments
-- but makes no other, substantive changes to the code. Will save that
for a rainy day.