Support submodules
Now you can run a submodule. The REPL is in a namespace for the module
body.

Caveat: There is a bug in Racket 6.2.1 and earlier (but fixed on HEAD).
        When errortrace is being used, module->namespace does not return
        the definitions in a module+ form. The module is evaluated
        fine (e.g. tests in a test submodule are run), but the
        definitions cannot be accessed in the REPL. (Plain `module` and
        non-#f `module*` forms do work fine in 6.2.1 and earlier, and
        when errortrace isn't used.) As a result, display warning/advice
        the first time a user does this each session.

- ,run also accepts a list (<file> <mod> ...)

- racket-run runs the innermost submodule enclosing point.

- racket-test is now ,run `(,(buffer-file-name) test) as opposed to
  squirting a require into the REPL. Compared to before, there are two
  benefits:

  - The prompt doesn't appear until the tests complete.

  - The user can experiment with tests in the REPL.

The hairiest part turned out to be handling all the permutations of
string vs. symbol paths, paths vs. module-paths, absolute vs. relative
paths, and so on. A new `mod` struct and related functions manage this.