Avoid false test failures on Travis CI
Call accept-process-output frequently until we see the output we expect
or we timeout.

Specifically this helps with the test that looks for the REPL banner and
prompt. Probably this had been failing intermittently due to output
buffering plus the fact that the banner and prompt are on separate lines.

Although that was the only unreliable test, so far, we may as well do
this generally for all such tests in `racket-test/see-rx'.