Simplify avoiding comint-highlight-input
Instead of a copypasta of comint-send-input minus the expression to add
'font-lock-face 'comint-highlight-input -- simply call comint-send-input
then immediately remove the property afterwards. Although I see it
briefly redisplay with the highlight when stepping in edebug, I don't in
normal use. Even if it were to flash momentarily, that wouldn't
necessarily look bad AFAICT.