Optimize finding module-level forms
New `racket--module-level-form-start' helper works much more quickly to
do the sort of identification needed for beginning-of-defun-function,
font-lock-extend-region-function, and syntax-begin-function.

The old approach could be extremely slow if a large string had many
chars with ( and ) syntax. We want to ignore such chars and jump
directly out of the string. (Although this will turn out to be important
for adding here-strings, especially because they tend to be
long/multiline strings, it's the same issue when expressing the same
string using plain old double quotes.)