Support Racket here-strings. Closes #176
Mainly borrow the approach used by sh-mode for here-docs.

By giving these actual string syntax things like indentation work as
desired (the same as for a double quoted multi-line string).

A custom font-lock-syntactic-face-function handles nearly all of the
font-lock, with pattern-based font-lock handling just the #<<ID portion.

The trickiest part is making the syntax-propertize function recognize both
the beginning and end of a here-string -- including recognizing the end of
a here-string whose beginning was identified in a previous call to the
syntax-propertize function (propertiziation happens in batches/windows,
the boundary of which may intersect a her-string).