Fix performance issues. Fixes #178
racket--module-level-form-start should not try to move out of strings or
comments. (Attempting to do so is too slow, because syntax-ppss scans
from the start of the buffer. Although syntax-ppss uses a cache, that is
invalidated when the buffer changes.)

Move that functionality to a new racket--escape-string-or-comment
function. Use it only from select interactive commands like
racket-beginning-of-defun-function and racket-backward-up-list. And,
don't use those interactive commands for non-interactive purposes like
font-lock region extension.

Where possible, don't even use racket--escape-string-or-comment, instead
use parse-partial-sexp. Or at least use racket--escape-string-or-comment
just once, before a series of backward-up-list or similar scans.

In the same spirit, optimize racket-smart-open-bracket to detect early
on whether point is in a string or comment.

Don't set syntax-begin-function, at all. Originally this had been set to
beginning-of-defun, and more recently commit a4f6e17 had changed this
racket--module-level-form-start. But the former was "circular" -- it was
using syntax-ppss to find a string or comment, whereas the point of
syntax-begin-function is to help syntax-ppps do this. And the latter is
now a no-op, as it neither moves point nor escapes strings/comments. It
turns out that doing nothing at all is just fine. We don't need this.