Add Racket 6.3 to matrix; remove some others
The total build matrix here is very large. Let's remove the older x.x.1
releases. Why those? No great reason. Just the admittedly questionnable
rationale that a problem is less likely to show up in a point release.

Of course when the latest stable release is, for awhile, a point
release, we'll include that.