Don't set compile-command; fixes #202
Although it would be possible to fix #202 via something like PR #204, in
hindsight I think automagically setting `compile-command` is a bad idea.
Even with this fix, it's still not quite right -- e.g. if the buffer's
file name changes. This is something a user may configure themselves
with racket-mode-hook, to work as pragmatically or thoroughly as they
prefer.

See also commit b0d01f8