racket--repl-command: Preserve point
Fixes #235.

When checking for an sexpr in the command response, we had been leaving
point moved to point-min. This could cause problems in edge cases,
particularly involving very long responses like that for the `syms`
command.

It would have been better to use save-excursion when moving point and
calling forward-list. But even better: Use scan-lists directly.