Call syntax-propertize before hook; fixes #222
`racket-syntax-propertize-function` handles things that a simple
syntax-table can't. But it might not run before hook functions run. As
a result, `paredit-mode`'s call to `check-parens` might fail.

To avoid such problems, do `(syntax-propertize (point-max))` early in
our initialization before hook funcs are called.

(I considered making this conditional on `paredit-mode` being in the
list of hook functions. However I found that the eager/full
`syntax-propertize` is quite fast, even on an 18,000 line file. And
`paredit-mode` might not be the only thing that depends on "exotic"
sexps already having correct syntax. So just go ahead and do it
always.)

Also, while the hood is open:

- Rename the common init function to `racket--common-variables` and
  move a few things among groups within it.

- A few other minor edits, such as editing comments and deleting a
  `declare` that is no longer needed.