Adds `racket-run-main` to run the `main` submodule
`racket-run-main` will run the `main` submodule of the current
submodule (respecting deep nesting), or the current submodule if it is
a `main` submodule (so if you hit `racket-run-main` when you're in a
`main` module, the command does what you'd expect).

This also adds the "C-c C-m" (or "C-c RET") key binding for
`racket-run-main`.

This commit also changes `racket-test` to run the `test` submodule of
the current module (so you can run it within deeply nested
submodules), or the current module if it is a `test` submodule.