racket-visit-{definition module}: Minimize "run?" prompts
Before: Would try to find the definition, and if not found, prompt to
run. Often running again wouldn't help (the definition simply can't be
found). For users it can feel like being stuck in a loop, not knowing
when to answer "n" and stop trying. See issue #288 discussion.

After: Before trying to find, offer to run IFF it would help (the REPL
doesn't have the buffer's contents). Then try to find. If not found,
just say so.

Implementation extends the "prompt" command to return md5 digest of
the file as well as the path. The front-end Elisp can use this to
determine if a run is needed.