Use file/md5 instead of openssl/md5
Although file/md5 is slower, it's not _that_ much slower, and .rkt
file sizes are KBs not MBs or GBs.

Furthermore, file/md5 is available in Racket 6.0 base, unlike
openssl/md5 which appears in 6.1. Maybe someday we'll bump the
requirements to 6.1 or even newer. But it doesn't need to be today,
just for this.