Add a cache of expanded syntax.
A "before run" function clears the cache.

An eval-handler populates the cache when it can. It includes
non-compiled files required by the file being run -- which is good, as
they're likely to be targets of e.g. find-definition. Also, they need
to be expanded anyway, so there is no meaningful speed cost to saving
them (just some memory cost).

An "after run" function handles the case where a zo was used.

Use new file->expanded-syntax, which is cache aware, in find-definition.