Make `run` command response wait until entering REPL
This will let Emacs use the command response as a reliable way to run
another command that needs the run to have completed. And by using the
async callback, to do so without blocking Emacs.

Specifically it makes it possible to easily compose `run` with a
subsequet command like `get-profile` or `get-uncovered`.

(As opposed to the old way of retrying in a loop until the `prompt`
command returns the value we want. Not only does this block Emacs, it
isn't even reliable in some cases like racket-test RE-running the
same module with coverage.)