Call async callbacks with run-at-time not directly
In trying to use the callback with racket-test to insert coverage
overlays from the callback, I observed this worked fine in Emacs 25.1
but not in 26.1. Although I didn't figure out precisely why, it
occurred to me it's probably better not to do the callback in the
dynamic extent of a process filter.