Better tests for racket-check-syntax-mode and generally
defconst a reasonable timeout to use for most things: 30 seconds
locally, 15 minutes on Travis CI.

Replace use of cl-some and a list of nils -- a great nominee for my
Srsly What Was I Thinking all-time Wall of Shame -- with a simple use
`with-timeout`.

Now that racket-check-syntax-mode completes asynchronously, we need to
wait check for that. Look for its non-nil header-line-format.