Have racket--repl-command-async handle error responses
That way, the callback only gets the result value from an 'ok
response. This simplifies callbacks that don't need to handle the
error response explicitly. Furthermore, it's better not to handle them
by calling `error' -- a simple `message' suffices and avoids a
confusing "error running timer" prefix to the error message.

The old behavior is available as `racket--repl-command-async-raw' for
when something needs to handle error responses explicitly.