Allow multiple commands in flight
Yet another change to command processing. Maybe the last, for awhile?

Previously, a command request had to have a response before another
command request would be processed.

In a scenario like doing racket-expand-file on a very large file, it
can take very many seconds until the response for the initial
expansion step. Because racket-expand-file uses an async command, the
Emacs UI as a whole is not blocked. But further Racket commands are.

One way to solve this would be for the command server to accept
multiple simultaneous connections. Each connection would handle a
request and response. Like HTTP. Although that would work, making a
connection for each request seems unnecessarily "heavy".

Another way to solve this is for command requests to supply a unique
ID (a "nonce", which can simply be an incrementing number). The
command server spins off a new thread to process each request. When
done, the nonce+response is put to a channel. A responder thread gets
these one by one from the channel and writes them one by one. Then on
the Emacs send, we simply keep a hash-table of nonces to callback
lambdas.

The second way is what's done in this commit.

Also take the opportunity to give some Emacs Lisp functions shorter
and more descriptive names.