Replace prompt and path commands with path+md5; remove at-prompt
Part of the motivation for the prompt command was to use `at-prompt`
to detect whether we're at a prompt. However, that was mostly used as
a kludgy way for things to see if a run command had finished before
issuing another command like get-profile or get-uncovered. Now that we
have command response callbacks, and are using them instead, the
at-prompt detection is an unncessary complication. Rip it out.