Handle multiple successive connections
Make the Elisp correctly detect when the connection needs to be
re-made.

Handle state where connecting but not yet connected.

In the Racket back-end, handle exceptions and use a custodian for each
connection (so that it's resilient in the face of disconnections).

Also: Rewrite using named functions instead of thunks -- and
better-named functions -- to hopefully improve clarity and decrease
nesting.

Also: Improve the comment for start-command-server.