Use accept-process-output not sit-for; fixes #318
Although sit-for will allow process output to be processed, it does
extra work that we don't need -- and which interferes with
company-mode.

(Reading the source code for sit-for, it uses read-event, and comments
refer to mailing list discussions about a messy history. It seems wise
to avoid sit-for unless the objective really is to redisplay and wait
until some user input arrives or time elapses. That is not the
objective here.)