Simplify completion-table-dynamic expression
While investigating issue #318, I became suspicious of this code:
Might it do buffer-substring-no-properties on the wrong buffer, if
company-mode changed current-buffer?

Maybe not; #318 turned out to have a different cause. But the code was
still silly: The function argument we were ignoring supplies the
prefix string; there is no need to get it from the buffer.