elisp->racket: Handle propertized strings; fixes #319
Instead of relying on the Emacs front end to de-propertize strings,
enhance the back end to recognize them and extract the plain string
portion.

Handling this centrally should prevent a class of bugs that tends to
crop up.

This relies on the read representation of Emacs propertized strings
being the same as Racket vectors: #().