Experimental step debugger
This arises from me taking a fresh look at the years-old debugger
branch that I never merged.

Compared to back then, now that the command mechanism has changed,
it's simpler to connect the front and back ends. We have a thread that
reads command responses from a channel and writes them as sexprs to
the TCP port. We simply extend this to also sync on a
debug-break-channel and write those, too. Up in Emacs, we look for
these (debug-break ___) sexprs -- easy to distinguish from (ok __)
and (error __) sexprs that are command responses -- and dispatch them
directly to a dedicated singleton callback function.

Like DrRacket this allows entire-file debugging. Unlike DrRacket:

- You're not prompted you about each file interactively; instead it
  uses the Emacs Lisp variable `racket-debuggable-files'.

- `racket-repl-mode' provides a debug prompt at each step, in which
  locals are available to inspect and even set!.

- You can do an edebug style C-u C-M-x to debug instrument individual
  function definitions.

A new `racket-debug-mode' minor mode provides visual feedback and
resume commands in the `racket-mode' buffer where the break happened.

I did need to "fork" gui-debugger/annotate to resolve two issues
I reported there:

  https://github.com/racket/drracket/issues/230
  https://github.com/racket/drracket/issues/231

Even if those are fixed, there, racket-mode is supposed to support
Racket versions as old as 6.0, so I think my fork will need to remain.