Better way to say what files to debug instrument
defvar `racket-debuggable-files`, which can be a list of files to
debug, or a function that returns such a list. Users can supply their
own function, or set it as a file-local var, etc.

Set it to a reasonable default, `racket-project-files`, which uses
projectile if available else returns files in the same directory as
the file being run.