Use yield not sync in prompt-read when racket/gui; fixes #326
The prompt-read handler runs on the REPL thread, which is the same
thing as the eventspace handler thread when racket/gui is loaded.
Now that we're blocking on a read-interaction channel, the
current-get-interaction-input-port can't do its magic because it's
running on the read-interaction thread. As a result, we need to call
yield from the REPL thread.