Move details to new interactions.rkt
Instead of directly exposing the channel, and requiring the
prompt-read handler to know how to read -- including the sync/yield
business -- hide that and simply expose:

- A get-interaction function for a prompt-read handler to call.

- A current-yield/sync parameter for the main thread to set to either
  sync (no GUI) or yield (GUI).

This also sets the stage better for a debugger REPL, which I'm working
on in another branch. Getting this on master will simplify rebasing to
master.