Add defcustom racket-retry-as-skeleton
Allow people to disable retrying a "skeleton" of a file with errors.

(This isn't in response to any specific complaint. I have noticed
that, when a file takes a noticeably long time to expand, this will of
course roughly double that time. Although the "best" answer to a
project with such a situation is probably, "Use raco make or raco setup
to make compiled bytecode!", let's provide a way to disable this retry
feature.)