Duplicate current-{input output error}-port; fixes #381
That way, a port-{read write display print}-handler set by a user
program will not persist into the next racket-run.

Another way to fix this might be to restore the original 4 handlers.
However the dup-{input output}-port docs mention this specific case.
Also it seems like duping the ports freshly for each run would address
any other such per-port state, which I'm not aware of (or which might
be added someday). So this seems like the best approach.