Change how racket-smart-open-bracket is initialized; fixes #385
Old: We always bound this, by default, and used a boolean variable
racket-smart-open-bracket-enabled to control whether it did the
special behavior. In an attempt to make this work automagically with
paredit, we had to fuss with the paredit-mode keymap in an
eval-after-load form. But apparently that can cause load-order
problems using the use-package :hook option. Keep in mind that
use-package :init forms also run in an eval-after-load form. In what
order these each run, isn't necessarily well-defined.

New: Don't try to be so clever and automagical. If people want to use
racket-smart-open-bracket, then they should do so the typical Emacs
way -- simply bind that command to the [ key in the racket-mode-map
and/or racket-repl-mode-map keymaps. If they are also using things
like paredit, then they can configure them however and whenever works
best; they are in control.