Fix bad binding for C-S-backspace; closes #387
Thanks to @wouteroostervld for reporting this problem by way of
submitting pull request #387 as a possible fix.

The root cause is the buggy "[C-S-backspace]" binding which ought to
be "<C-S-backspace>". The former was effectively making "[" -- and "C"
and "S" and so on -- prefix keys! Ugh.

This bad binding had been hidden by another, explicit binding of "[" to
racket-smart-open-bracket until commit 669275b a couple weeks ago. (It
is also hidden if you use paredit-mode, which has an explicit binding
of "[").