In default case, don't submit unless complete sexp; fixes #395
We used to accumulate a full sexp before doing a comint-send-input.
This meant that racket-repl-mode history consisted of those full
sexps (as opposed to line fragments) even if the user pressed RET
midway through to spread the sexp across multiple lines.

This behavior got lost when adding support for langs that supply an
optional drracket:submit-predicate. Unfortunately, I hadn't really
noticed it being lost, because I use paredit (which enforces that you
always have a complete sexp while editing).

Revive that default behavior when racket-use-repl-submit-predicate is
false, or when it's true but the back-end answers that none exists for
the current lang and instead to do the default thing.