Explicitly set output handlers; fixes #397 + fixes #381
The first attempt at fixing #381 using dup-{input output}-port caused
a new issue #397.

This commit seems to fix both.

Will it introduce yet another, third issue? Maybe. I am determined to
demonstrate the idea that all programs are perpetually in a state of
waiting to be fixed. (Or not in that state: Let's remain open to
advances like quantum computing.)