Always do racket--repl-before-run-hook; fixes #406
Effectively this reverts a portion of commit 2b1c7d4 which seems
unrelated to the rest of the commit and unjustified. The debugger
didn't add a before-run hook -- much less one that would cause a
problem without a live REPL process. So, what was I thinking??

It's possible now I will regress some bug. I hope not, but if so and
the change needs to be "re-reverted"? At least it would be its own
small commit with a clear justification. Anyway, in that case keeping
issue 406 fixed would mean distinguishing before-run hooks that should
always run vs. should only run if live REPL.