WIP: Make racket-check-syntax-mode read/write
This way, people can leave this minor mode enabled for enhanced
highlighting and information.

As of this commit, the check-syntax annotation is only refreshed from
an after-save-hook. Not sure if this should instead be done "all the
time" from after-change-hook, maybe with a short delay timer. In that
case, we'd probably want to change the back-end command to accept the
buffer text instead of a file name. That in turn would allow fixing an
annoyance with check-syntax-mode-rename, which doesn't itself refresh
annotations (you must save).