WIP racket-check-syntax-mode
Change the back end command to take the program text, so that it can
work regardless of whether a buffer is saved to a file.

Refresh annotations in an after-change-hook with an idle timer delay,
instead of an after-save hook.

Comment-out the thing where we use an overlay 'after-string to show
the help-echo. I found it makes text jump around too much to be used while
editing.

Provide the best of both worlds for visiting definitions. For
non-imports, use the check-syntax definition (which works even for
shadowed bindings in internal definition context). Otherwise, use the
normal racket-visit-definition (which takes you to the definition in
the other file, instead of just plopping you on a require or #lang in
the current file).