WIP check-syntax: Only return local defs/uses
This simplifies things like the rename and visit commands, which want
to ignore imports from require or module lang.

It also simplifies the visual experience for the user. e.g. When you
put point on a `define`, do you _really_ need/want to see it
highlighted -- along with _everything_ from racket or racket/base?
Nah.

Also simplify some other code.

Still to-do in next commit: Improve the situation where you edit the
file but there is some syntax error. Currently we retain the
annotations (because they may be useful while fixing the error). But,
the error itself might be incorrectly highlighted. Ideally, we could
add some kind of error annotation at that site, in red, with the
help-echo being the error message.