WIP check-syntax: Supply more information
Be more ambitious: Supply more information and commands in racket-mode
edit buffers -- and do so without requiring racket-run.

- Supply more information from drracket/check-syntax, including
  annotations for documentation and the source files of definitions.

- Supply some information that drracket/check-syntax does not --
  completion candidates from imported modules.

- Use this in the front end racket-check-syntax-mode to augment commands
  like racket-doc, racket-visit-definition, and racket-complete.

In fact, this is moving to where `racket-mode` edit buffers need not
racket-run to get most of the useful information. (The back end server
does need to be started -- but not necessarily racket-run-ing the edit
buffer's file.) I think this would be a good improvement.

[Having said all that, `racket-repl-mode` will need to -- and ought to
-- continue using the status quo techniques that depend on having a
live module->namespace.]

The hardest part of this was determining the completion candidates for
imports, by parsing #%require grammar and using module->exports. (We
can't use module->namespace and namespace-mapped-symbols. We don't
want to use any other method that inadvertently instantiates the
module -- which would effectively run the user's program, including
side-effects!)

Caveat: There are still quite a few TODO and FIXME comments.