WIP check-syntax: 'namespace vs path-string commands
Some commands now have a first argument, "how" (a.k.a. "context"),
which is either 'namespace or a path-string.

So far this includes def and describe. Probably will also include doc.

Given some string representation of an identifier, this says how it
should be made into an actual identifier.

The 'namespace choice means to use namespace-symbol->identifier. This
is suitable for commands running in racket-repl-mode, and uses the
module->namespace namespace from the last racket-run. We let people
racket-run inside a specific submodule, which is the module->namespace
used then, so that works fine. Essentially this was the status quo,
only approach available before.

The path-string choice means, make the identifier by giving it the
lexical context from the 'module-body-context syntax property of an
expanded module form. This is suitable for commands running in
racket-mode enhanced by racket-check-syntax-mode. At least, it works
for identifiers at the module level; it doesn't know about submodules.
As a result, when we have annotated specific buffer text from
syncheck:add-jump-to-definition and syncheck:add-docs-menu, we use
that information because it will be accurate. But when the user does a
C-u flavor of a command to enter something at the prompt, or, for
things like :company-location or :company-doc-buffer, we have to just
assume module-level -- that's where some commands with a path-string /
module-body-context flavor are useful.

As part of this, I exploded the old identifier-binding* function into
useful pieces, and moved that to its own new file, identifier.rkt.

Also add a couple loggers, racket-mode-check-syntax and
racket-mode-check-syntax-cache. These work fine with
racket-logger-mode so that's handy when using Racket Mode to develop
Racket Mode. :)