WIP check-syntax: Improve racket-check-syntax-describe
When it has a racket-check-syntax-doc property, supply its path and
anchor values to the back-end "describe" command. In this case, no
xref-binding->definition-tag is needed -- or wanted, it might be
wrong! -- and so we use "describe" solely for its ability to extract
and clean the HTML for a shr buffer.

Also, update the doc string for racket-check-syntax-mode in light of
recent changes.